ibge/requirements.txt
Django>=1.4
django-tastypie>=0.9.11
